Rename math node types file
