+++ src/models/mathNodeTypes.ts
